modifié :         ifs-cars-server/src/main/java/fr/uge/corp/ifscars/cars/Car.java 	modifié :         ifs-cars-server/src/main/java/fr/uge/corp/ifscars/cars/ICar.java 	modifié :         ifs-cars-server/src/main/java/fr/uge/corp/ifscars/cars/Storage.java 	modifié :         ifs-cars-server/src/main/java/fr/uge/corp/ifscars/command/Command.java 	modifié :         ifs-cars-server/src/main/java/fr/uge/corp/ifscars/rating/IRating.java 	modifié :         ifs-cars-server/src/main/java/fr/uge/corp/ifscars/renting/IRentingService.java 	modifié :         ifs-cars-server/src/main/java/fr/uge/corp/ifscars/renting/RentingService.java 	modifié :         renting-app/src/main/java/fr/uge/corp/ifscars/cars/ICar.java 	modifié :         renting-app/src/main/java/fr/uge/corp/ifscars/rating/IRating.java 	modifié :         renting-app/src/main/java/fr/uge/corp/ifscars/rating/Rating.java 	modifié :         renting-app/src/main/java/fr/uge/corp/ifscars/renting/IRentingService.java 	modifié :         renting-app/src/main/java/fr/uge/corp/rentingapp/command/Command.java

+++ renting-app/src/main/java/fr/uge/corp/ifscars/cars/ICar.java
@@ -21,7 +21,7 @@ public interface ICar extends Remote {
 	/**
 	 * @return Getter for price.
 	 */
-	double getPrice() throws RemoteException;
+	long getPrice() throws RemoteException;
 	
 	/**
 	 * @return Getter for id.
